requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
python-dateutil==2.8.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
sentence-transformers==2.2.2
numpy==1.24.3
werkzeug==3.0.1

